CPU 수정 / 첫번째 실행 완료

+++ Apple2/AppleMachine.h
@@ -10,7 +10,7 @@ void machine_InitMachine();
 bool machine_Booting();
 
 void machine_Reset();
-void machine_Run(int cycle);
+void machine_Run(unsigned long cycle);
 void machine_Render(int frame);
